requirements.txt
streamlit
pandas
numpy
faiss-cpu
PyPDF2
openai
